déplace un composant partagé
