+++ app.territoiresentransitions.react/src/app/pages/collectivite/Indicateurs/detail/IndicateurDetail.tsx
@@ -1,7 +1,7 @@
 import {BadgeACompleter} from 'ui/shared/Badge/BadgeACompleter';
 import {Badge} from 'ui/shared/Badge';
 import {referentielToName} from 'app/labels';
-import {ActionsLieesCards} from '../../PlansActions/FicheAction/FicheActionForm/ActionsLiees';
+import {ActionsLieesCards} from '../../PlansActions/FicheAction/FicheActionForm/ActionsLieesCards';
 import IndicateurChart from '../charts/IndicateurChart';
 import {IndicateurValuesTabs} from './IndicateurValuesTabs';
 import {TIndicateurReferentielDefinition} from '../types';
