+++ app.territoiresentransitions.react/src/app/pages/collectivite/Indicateurs/detail/IndicateurEnfant.tsx
@@ -6,7 +6,7 @@ import {BadgeACompleter} from 'ui/shared/Badge/BadgeACompleter';
 import {useToggle} from 'ui/shared/useToggle';
 import {ExpandToggle} from 'ui/icons/ExpandToggle';
 import {useIndicateurACompleter} from '../useIndicateurDefinitions';
-import {ActionsLieesCards} from '../../PlansActions/FicheAction/FicheActionForm/ActionsLiees';
+import {ActionsLieesCards} from '../../PlansActions/FicheAction/FicheActionForm/ActionsLieesCards';
 import IndicateurChart from '../charts/IndicateurChart';
 import {FichesActionLiees} from '../FichesActionLiees';
 import {Spacer} from 'ui/dividers/Spacer';
